+++ src/AliasNameBody.ts
@@ -1,0 +1,6 @@
+
+export interface AliasNameBody {
+  aliasName: string;
+  aliasBody: string;
+  description: string | undefined;
+};